GoogleTranslateClient/translateclient/requirements.txt
functions-framework==3.5.0
google-cloud-translate==2.0.1